learny/microblog/requirements-flask.txt
appdirs>=1.4.3
blis>=0.7.4
certifi>=2019.6.16
chardet>=3.0.4
Click>=8.0.1
cymem>=2.0.5
#de-core-news-sm>=2.1.0
Flask>=1.1.1
idna>=2.8
itsdangerous>=1.1.0
jinja2>=2.11.3
lxml>=4.6.3
MarkupSafe>=1.1.1
murmurhash>=1.0.2
#nltk>=3.4.5
#numpy>=1.17.0
plac>=0.9.6
preshed>=3.0.5
PyHyphen
python-docx>=0.8.11
requests
six>=1.16.0
spacy>=3.1.0
srsly>=2.4.1
textblob>=0.15.3
textblob-de>=0.4.3
thinc>=8.0.7
tqdm>=4.61.2
urllib3>=1.26.5
wasabi>=0.8.2
Werkzeug>=2.0.1
